refactor(compositions): créer composant DiscordMessageEditor avec MentionSuggestions
- Extraire l'éditeur de message Discord avec autocomplete des mentions
- Déplacer MentionSuggestions dans DiscordMessageEditor
- Supprimer les états mentionAnchor et mentionQuery du hook useDiscordMessage
- Simplifier la structure de page.tsx
- Réduire encore la taille du fichier (2662 -> 2250 lignes, -412 lignes)
